requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
Django==4.0.5
django-ckeditor==6.5.0
django-filter==22.1
django-js-asset==2.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
docopt==0.6.2
google==3.0.0
google-api-core==2.10.0
google-api-python-client==2.61.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.4
Markdown==3.3.7
oauth2client==4.1.3
oauthlib==3.2.1
Pillow==9.2.0
pipreqs==0.4.11
protobuf==4.21.5
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
yarg==0.1.9
zipp==3.8.0
